+++ tgstation.dme
@@ -336,7 +336,6 @@
 #include "code\__HELPERS\icon_smoothing.dm"
 #include "code\__HELPERS\icons.dm"
 #include "code\__HELPERS\jatum.dm"
-#include "code\__HELPERS\jobs.dm"
 #include "code\__HELPERS\level_traits.dm"
 #include "code\__HELPERS\levels.dm"
 #include "code\__HELPERS\lighting.dm"
