+++ tgui/packages/tgui/interfaces/common/JobToIcon.ts
@@ -1,0 +1,6 @@
+// Used to handle jobs that don't have a trim.
+export const JobToIcon = {
+  'AI': 'eye',
+  'Cyborg': 'robot',
+  'Personal AI': 'mobile-alt',
+} as const;
